requirements.txt
black==25.1.0
blinker==1.9.0
click==8.2.0
Flask==3.1.1
flask-cors==5.0.1
flask-marshmallow==1.3.0
flask-swagger-ui==5.21.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
python-dotenv==1.1.0
Werkzeug==3.1.3
